+++ package.json
@@ -17,18 +17,15 @@
     "ent": "^2.2.0",
     "express": "4.15.2",
     "express-socket.io": "^1.0.3",
-    "farmhash": "^2.0.4",
     "fcm-push": "latest",
     "forever-monitor": "^1.7.1",
     "ioredis": "^3.2.2",
     "json-align": "^0.1.0",
     "logjsx": "0.0.2",
     "node-json-db": "^0.7.3",
+    "redis": "^2.8.0",
     "rus-diff": "^1.1.0",
     "sha1": "^1.1.1",
-    "socket.io": "^2.0.4",
-    "socket.io-redis": "^5.2.0",
-    "sticky-session": "^1.1.2",
     "sync-node": ">= 1.3.1"
   },
   "keywords": [
